purealpha_mvp/requirements.txt
# PureAlpha 3.1 MVP - Python Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Science & ML
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.0
statsmodels==0.14.1

# Data Sources
yfinance==0.2.35
pandas-datareader==0.10.0
requests==2.31.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Caching & Queue
redis==5.0.1
celery==5.3.6

# Visualization
streamlit==1.30.0
plotly==5.18.0
matplotlib==3.8.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Logging
loguru==0.7.2

# Real Data Sources (Phase 1)
fredapi==0.5.1
pandas-datareader==0.10.0

# Content Generation & Auto-posting (Phase 2)
anthropic==0.39.0
python-telegram-bot==20.7
linkedin-api==2.3.0
schedule==1.2.0
